% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpTtestTable}
\alias{cpTtestTable}
\title{Care Partners Results Table}
\format{
##\code{cpTtestTable}
a data frame with 61 rows and 8 columns
\describe{
\item{Instrument}{Instrument ID}
\item{dimension}{dimension name, if applicable}
\item{Mean Difference}{Mean Difference in score from PreTest to PostTest}
\item{Std.Err.}{Standard Error of mean difference}
\item{95\% CI}{95\% Confidence Interval of mean difference in score from PreTest to PostTest}
\item{p-value}{pvalue from t test}
\item{t}{statistic from t test}
\item{df}{degrees of freedom from ttest}
}
}
\usage{
cpTtestTable
}
\description{
Care Partners Results Table
}
\keyword{datasets}
