% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demdata1}
\alias{demdata1}
\title{Demographics Data Part 1}
\format{
##\code{demdata1}
A data frame with 17 rows and 5 columns
\describe{
\item{RecordID}{Participant ID}
\item{Role}{Care Partnet or PD Patient}
\item{Gender}{Male or Female}
\item{Years of Education}{Number of Years of Schooling the participant completed, starting in elementary and going through any college. 12 years of education is equivalent to completing a high school education.
}
\item{Age}{Age in Years at time of beginning of tandem cycling intervention.}
}
}
\usage{
demdata1
}
\description{
Demographics Data Part 1
}
\keyword{datasets}
