% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demdata2}
\alias{demdata2}
\title{Demographics Data Part 2}
\format{
##\code{demdata2}
A data frame with 9 rows and 6 columns
\describe{
\item{RecordID}{Participant ID}
\item{Age at Dx}{Age at time of Parkinson's Disease Diagnosis}
\item{HoenandYahr}{Hoen and Yahr Score}
\item{CD}{Prescribed dosage of Carbidopa in mg/day}
\item{LD}{Prescribed dosage of Levodopa in mg/day}
}
}
\usage{
demdata2
}
\description{
Parkinson's Disease information, including staging and prescribed medications.
Not applicable for Care Partners.
}
\keyword{datasets}
