% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demographicsTable}
\alias{demographicsTable}
\title{Demographics Data in a Table ready to put in a document}
\format{
##\code{demographicsTable}
a data frame with 12 rows and 5 columns
\describe{
\item{Role}{Care Partner or Person with Parksinson's Disease}
\item{var}{Variable}
\item{meansd}{mean and standard deviation formatted in character format as 'mean (sd)'}
\item{mediqr}{median and IQR formatted as characters like 'median (IQR)'}
\item{level}{level}
}
}
\usage{
demographicsTable
}
\description{
Demographics of study participants
}
\keyword{datasets}
