% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dimension_meansd}
\alias{dimension_meansd}
\title{Mean and Standard Deviations of Pre and Post Test Survey Dimension Score}
\format{
##\code{dimension_meansd}
A data frame with
315 rows and 5 columns
\describe{
\item{Role}{Care Partner or PD Patient}
\item{dimension}{Survey Dimension}
\item{survey}{Survey}
\item{Test}{Pre Test or Post Test}
\item{mean}{Mean}
\item{sd}{Standard Deviation}
}
}
\usage{
dimension_meansd
}
\description{
Mean and Standard Deviations of Pre and Post Test Survey Dimension Score
}
\keyword{datasets}
