% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dimension_ttests}
\alias{dimension_ttests}
\title{t-test results}
\format{
##\code{dimension_ttests}
A data frame with 105 rows and 10 columns
\describe{
\item{Role}{Care Partner or PD Patient}
\item{Instrument}{Instrument ID}
\item{dimension}{survey dimension}
\item{mean}{mean}
\item{se}{Standard Error}
\item{conflow}{Lower Limit of 95\% Confidence Interval}
\item{confhigh}{Upper Limit of 95\% Confidence Interval}
\item{p}{P value from Test}
\item{t}{t statistic}
\item{df}{Degrees of Freedom}
}
}
\usage{
dimension_ttests
}
\description{
t-test results
}
\keyword{datasets}
