% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{do_score}
\alias{do_score}
\title{Score all dimensions}
\usage{
do_score(splitted, dimname, ...)
}
\arguments{
\item{splitted}{data frame split by survey}

\item{dimname}{name of the dimension}

\item{...}{additional parameters}
}
\value{
named vector with the following values:
RecordID = character,
Role = character,
Test = character,
survey = character,
dimension = character,
score = numeric
}
\description{
Score all dimensions
}
