% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fb_writemerge}
\alias{fb_writemerge}
\title{merge firstbeat files of the same id and test}
\usage{
fb_writemerge(df, id, t, dest)
}
\arguments{
\item{df}{data frame of filepaths, ids, and tests}

\item{id}{id to filter by}

\item{t}{test to filter by}

\item{dest}{file destination}
}
\value{
data frame
}
\description{
merge firstbeat files of the same id and test
}
