% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_firstbeat.R
\name{filter_firstbeats}
\alias{filter_firstbeats}
\title{filter by recordID and Test type}
\usage{
filter_firstbeats(id, test, df)
}
\arguments{
\item{id}{participant id}

\item{test}{pre test or post test}

\item{df}{dataframe of id, test, and filenames}
}
\value{
list
}
\description{
filter by recordID and Test type
}
