% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_firstbeat}
\alias{get_firstbeat}
\title{get path to first beat raw data files}
\usage{
get_firstbeat(path = NULL)
}
\arguments{
\item{path}{optional specific file name}
}
\value{
character string of path to file
}
\description{
get path to first beat raw data files
}
