% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hrv24hr}
\alias{hrv24hr}
\title{24 hour Heart Rate Variability Data}
\format{
##\code{hrv24hour}
a data frame with 70 rows and 6 columns
\describe{
\item{RecordID}{Participant Study ID}
\item{variable}{Heart Rate Variabilty Variable}
\item{PostTest}{value at PostTest}
\item{PreTest}{value at PreTest}
\item{Difference}{Difference between PreTest and PostTest}
\item{Role}{Participant Role, Care Partner or Person with Parkinson's Disease}
}
}
\usage{
hrv24hr
}
\description{
Results from Kubios HRV Premium after analysing the
FirstBeat data from all participants
}
\keyword{datasets}
