% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_firstbeat.R
\name{insert_rr_interval}
\alias{insert_rr_interval}
\title{Insert an RR interval duration to join two data frames of rr data from a
First Beat Heart Rate sensor}
\usage{
insert_rr_interval(df1, df2)
}
\arguments{
\item{df1}{data frame of first section of RR interval data}

\item{df2}{data frame of second section of RR interval data}
}
\value{
data frame
}
\description{
Insert an RR interval duration to join two data frames of rr data from a
First Beat Heart Rate sensor
}
