% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{interpexp}
\alias{interpexp}
\title{create comparison expression for different values}
\usage{
interpexp(df, var)
}
\arguments{
\item{compare}{}
}
\value{
expression
}
\description{
create comparison expression for different values
}
