% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interpret_promis}
\alias{interpret_promis}
\title{Scored PROMIS dimensions}
\format{
##\code{interpret_promis}
A data frame with 357 rows and 5 columns
\describe{
\item{RecordID}{Participant ID}
\item{Role}{Care Partner or PD Patient}
\item{dimension}{PROMIS dimension}
\item{Test}{Pretest, Posttest, or Difference}
\item{Tscore}{Tscore of actual score on the dimension}
}
}
\usage{
interpret_promis
}
\description{
Scored PROMIS dimensions
}
\keyword{datasets}
