% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interpret_scores}
\alias{interpret_scores}
\title{Interpreted Scores for Surveys that use Interpretations}
\format{
##\code{interpret_scores}
A data frame with 170 rows and 6 columns
\describe{
\item{RecordID}{Participant ID}
\item{Role}{Care Partner or PD Patient}
\item{Test}{Pre Test or Post Test}
\item{survey}{survey id}
\item{score}{scored dimension}
\item{Interpretation}{Interpretation of score}
}
}
\usage{
interpret_scores
}
\description{
Interpreted Scores for Surveys that use Interpretations
}
\keyword{datasets}
