% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{map_dims}
\alias{map_dims}
\title{purrr::map across the dimensions}
\usage{
map_dims(splitted)
}
\arguments{
\item{splitted}{data frame split on survey}
}
\value{
data.table
}
\description{
purrr::map across the dimensions
}
