% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_firstbeat.R
\name{merge_firstbeat}
\alias{merge_firstbeat}
\title{Merge two seperate files from a FirstBeat Heart Rate sensor}
\usage{
merge_firstbeat(filenames)
}
\arguments{
\item{filenames}{list of filenames that hold the data to merge}
}
\value{
dataframe
}
\description{
Merge two seperate files from a FirstBeat Heart Rate sensor
}
\details{
When the FirstBeat is disconnected or turns off and then turns
back on, it creates a seperate file. The Kubios HRV Premium
software cannot take multiple files as input. In order to
analayse all the data together if the sensor created multiple
files during the time of measurement, the files have to be
merged by hand.

The FirstBeat measures RR intervals. In order to merge the files
together, the function creates an RR interval that is the
length of time between one measurement ending and the next
measurement beginning.
}
