% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pluckvalue}
\alias{pluckvalue}
\title{Score the answer based on the lookup table}
\usage{
pluckvalue(Survey, ItemNum, Answer, ...)
}
\arguments{
\item{Survey}{survey id}

\item{ItemNum}{item number}

\item{Answer}{answer given}

\item{...}{additional parameters}
}
\value{
scored value
}
\description{
Score the answer based on the lookup table
}
