% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pretty_confint}
\alias{pretty_confint}
\title{Get confidence intervals and pretty print them from data frame}
\usage{
pretty_confint(df, survey, role, dimname = NULL)
}
\arguments{
\item{df}{data frame with t test data}

\item{survey}{name of survey}

\item{role}{role}
}
\value{
string
}
\description{
Get confidence intervals and pretty print them from data frame
}
