% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pretty_meansd}
\alias{pretty_meansd}
\title{Calculate and Display mean and standard deviation of a vector}
\usage{
pretty_meansd(vec)
}
\arguments{
\item{vec}{numeric vector}
}
\value{
string of the format 'mean (sd)'
}
\description{
Calculate and Display mean and standard deviation of a vector
}
