% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pretty_print_val}
\alias{pretty_print_val}
\title{Select a parameter from a dataframe based on supplied conditions}
\usage{
pretty_print_val(df, cond, param)
}
\arguments{
\item{df}{dataframe of values to look in}

\item{cond}{condition to subset on}

\item{param}{parameter to return}
}
\value{
value
}
\description{
Select a parameter from a dataframe based on supplied conditions
}
