% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{promistscore}
\alias{promistscore}
\title{Score a dimension of the PROMIS}
\usage{
promistscore(score, dimname, df)
}
\arguments{
\item{score}{actual score}

\item{dimname}{name of the dimension}

\item{df}{dataframe of promis scoring info}
}
\value{
tscore
}
\description{
Score a dimension of the PROMIS
}
