% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_firstbeat.R
\name{read_firstbeat}
\alias{read_firstbeat}
\title{Read in First Beat data and assign an attribute for starttime}
\usage{
read_firstbeat(filename)
}
\arguments{
\item{filename}{name of file of First Beat data}
}
\value{
dataframe with attribute starttime
}
\description{
Read in First Beat data and assign an attribute for starttime
}
