% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_studydata}
\alias{read_studydata}
\title{Access Raw Data Files}
\usage{
read_studydata(path = NULL)
}
\arguments{
\item{path}{name of file to access}
}
\value{
file path
}
\description{
Access Raw Data Files
}
