% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{score_dim}
\alias{score_dim}
\title{Score the values using the appropriate method}
\usage{
score_dim(method, values)
}
\arguments{
\item{method}{scoring method}

\item{values}{values to score}
}
\value{
score
}
\description{
Score the values using the appropriate method
}
