% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scoreInterpretationTable}
\alias{scoreInterpretationTable}
\title{Score Interpretation of Instruments}
\format{
##\code{scoreInterpretationTable}
a data frame with 13 rows and 6 columns
\describe{
\item{Instrument}{Instrument ID}
\item{Interpretation}{Interpretation of Score}
\item{CP_PreTest}{Count of Care Partners who scored that Interpretation for PreTest}
\item{CP_PostTest}{Count of Care Partners who scored that Interpretation for PostTest}
\item{PwP_PreTest}{Count of Persons with Parkinson's who scored that Interpretation for PreTest}
\item{PwP_PostTest}{Count of Persons with Parkinson's who scored that Interpretation for PostTest}
}
}
\usage{
scoreInterpretationTable
}
\description{
Interpretation of participant's scores on instruments at
pre and post testing
}
\keyword{datasets}
