% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{select_eval}
\alias{select_eval}
\title{Select the appropriate scoring for each item on each survey}
\usage{
select_eval(Survey, ItemNum, ...)
}
\arguments{
\item{Survey}{survey id}

\item{ItemNum}{item number}

\item{...}{additional parameters}
}
\value{
string of the name of the lookup table to use for scoring this item
}
\description{
Select the appropriate scoring for each item on each survey
}
