% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_firstbeat.R
\name{str_date}
\alias{str_date}
\title{Get time from filenames}
\usage{
str_date(char)
}
\arguments{
\item{char}{character string of filename}
}
\value{
lubridate datetime object
}
\description{
Get time from filenames
}
