% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey_data}
\alias{survey_data}
\title{Survey Data from Tandem Cycling Parkinson's Disease Study}
\format{
\subsection{\code{survey_data}}{

A data frame with 4726 rows and 7 columns:
\describe{
\item{RecordID}{Participant ID}
\item{Role}{Care Partner or PD Patient}
\item{Test}{Pre Testing or Post Testing}
\item{Instrument}{Survey ID:
BC -Brief Cope,
BRS - Brief Resilience Scale,
GAD7 -General Anxiety Disorder 7 Item,
GD -Geriatric Depression Scale,
PDQ39 - Parkinson's Disease Questionnaire 29 Items,
PROMIS - PROMIS Global Health,
RDA- Revised Dyadic Adjustment Scale
}
\item{ItemNum}{Item Number of the survey}
\item{Answer}{Item Answer}
\item{Value}{Scored Value of the Answer}
}
}
}
\usage{
survey_data
}
\description{
Patients with Parkinson's Disease and their care partners completed this
8 week tandem cycling intervention. Patients and their care partners
completed these surveys before and after the intervention.
}
\details{
Data was recorded in RedCap and exported upon study completion.
The de-identified raw data from RedCap is in the 'data-raw' folder.
Cleaning and processing the data is documented in 'data-raw/SURVEYDATA.r'
}
\keyword{datasets}
