% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fga}
\alias{fga}
\title{Gait Analysis Data}
\format{
##\code{fga}
a data frame with 36 rows and 5 columns
\describe{
\item{RecordID}{Participant Study ID}
\item{Instrument}{Gait Analysis Instrument ID}
\item{PreTest}{Result at PreTest}
\item{PostTest}{Result at PostTest}
\item{Difference}{Difference between results from PreTest to PostTest}
}
}
\usage{
fga
}
\description{
Gait Analysis Data
}
\keyword{datasets}
